Create main.py
This was a experimental project didn't work alot on it as there are so many possibilities but they don't make me excited to do them.
This program takes some words and then use it in a story. It might sometimes make story with no sense. But is fun to read it.